Desktop/Hoarder/BE/requirements_dev.txt
asgiref==3.3.4
coverage==5.5
Django==3.2
django-heroku==0.3.1
djangorestframework==3.12.4
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1